Add files via upload
This project focuses on predicting online shoppers' intentions to make a purchase using the Online Shoppers Intention dataset. The goal is to uncover patterns and factors that influence user behavior on an e-commerce website and build a predictive model to identify potential customers.

The project utilizes the "Online Shoppers Intention" dataset, which includes various features such as session duration, pages visited, and user demographics. The dataset also contains the target variable indicating whether a user made a purchase or not